Concurrent first exercises
4.1. Exercise 1
The exercise asks to analyze what the two variants shown in the previous section do and study their nuances from an implementation point of view. To do this, you need to compile and run the programs in Listings 1 and 2 (files practica_1_V1.cpp and practica_1_V2.cpp from the provided materials, respectively), and after several executions, try to explain the behavior of the programs.

You should note that in line 11 of both programs, we are using the + operator for string concatenation: the result of "Soy " + nombre + "\n" is a string composed of the concatenation of "Soy ", the nombre parameter, and a newline at the end. In a sequential program, the instructions in lines 10 and 11 would behave the same way. However, is this true in the case of a concurrent program?

To answer this question, you can try running the program with those lines alternately commented and try to explain the differences in behavior.

Additionally, you should use standard documentation sources for the specification and usage of unknown elements, such as this_thread::sleep_for(...), chrono::milliseconds(...), etc.

4.2. Exercise 2
The task is to develop a new version of the program, following the pattern of the second approach shown, which will be called exercise_2.cpp, in which 10 processes are launched. The i-th process sleeps for a random time between 100 and 300 milliseconds and writes the corresponding message ("Soy 1", "Soy 2", ..., "Soy 10") a random number of times, between 5 and 15. To generate random numbers, you can refer to the documentation for the rand() and srand() functions in the cstdlib library.

For this exercise, we change the specification of the previous procedure to the following, where id is the identifier of the process that calls it (a value between 1 and 10):
void saludo(int id, int retardo, int veces)
4.3. Exercise 3
4.3.1. Exercise Description
The task is to create a program for data analysis using a concurrent program. You need to develop a concurrent program according to the following specifications:

Create and initialize a vector of 100 random real numbers.
Launch three processes to analyze the data: one that calculates the average, another that calculates the maximum and minimum values, and a third that calculates the standard deviation.
Once these values have been calculated, the main program outputs them to the standard output, formatted as follows:
makefile
# data: 100
average: some_value
max: some_value
min: some_value
sigma: some_value
The file containing the corresponding program for this exercise will be called exercise_3.cpp.